backend/requirements.txt
asgiref==3.7.2
Django==4.2.9
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
